==> text_analysis/3_mapping_rating_ids_kununu.ipynb <==
# Mapping Rating IDs to Categories in Kununu Reviews
-------------------

> <i>Description: In this notebook, we analyze and map rating IDs from Kununu reviews to specific categories, we use in our classification and analysis.</i>

We will use output excel file to add more specific keywords to our categories, calculate sentiment for textual data from rating text in Kununu reviews and map score ids into categories as well.

Input Files: 
1) reviews_merged.csv

Output:
1) Kununu_rating_ids_mapped.xlsx

import pandas as pd
import numpy as np
import re
from collections import Counter

* reviews_merged.csv is a result of merging Glassdoor and Kununu translated files. 

df = pd.read_csv('reviews_merged.csv')
print(df.head())
# To see what type of data is in ratings_translated
print(type(df['ratings_translated'].iloc[0]))

# Function to extract 'id' values from strings
def extract_ids_from_string(x):
    """
    Extracts 'id' values from a JSON-like string using regular expressions.

    Parameters:
    - x (str): A string containing nested JSON-like structures with 'id' key-value pairs.

    Returns:
    - list: A list of extracted 'id' values as strings. If the input is not a string, returns an empty list.
    """
    if isinstance(x, str):
        # Use a regex to find patterns like 'id': 'something'
        ids = re.findall(r"'id': '(\w+)'", x)
        return ids
    else:
        return []  # If it's not a string, return an empty list

# Applying the function to the 'ratings_translated' column
id_list = df['ratings_translated'].apply(extract_ids_from_string)

# Flattening the list of lists and count occurrences of each id
id_flat = [item for sublist in id_list for item in sublist]
id_counts = Counter(id_flat)

# Converting the counts to a DataFrame for better readability
id_counts_df = pd.DataFrame(id_counts.items(), columns=['id', 'count'])

print(id_counts_df)

# Mapping each id name into a category:
id_to_category = {
    "salary": "Financial Compensation & Benefits",
    "communication": "Leadership & Communication",
    "atmosphere": "Collaboration & Teamwork & Social Culture",
    "image": "Authenticity",
    "workLife": "Work-Life Balance & Flexibility",
    "career": "Professional Development and Continuous Learning",
    "environment": "Collaboration & Teamwork & Social Culture",
    "teamwork": "Collaboration & Teamwork & Social Culture",
    "oldColleagues": "Collaboration & Teamwork & Social Culture",
    "leadership": "Leadership & Communication",
    "workConditions": "Work-Life Balance & Flexibility",
    "equality": "Diversity & Equity & Inclusion",
    "tasks": "Professional Development and Continuous Learning",
    "apprenticeshiptasks": "Professional Development and Continuous Learning",
    "apprenticeshipatmosphere": "Collaboration & Teamwork & Social Culture",
    "apprenticeshipcareer": "Professional Development and Continuous Learning",
    "apprenticeshipworkHours": "Work-Life Balance & Flexibility",
    "apprenticeshipsalary": "Financial Compensation & Benefits",
    "apprenticeshipsupervisor": "Leadership & Communication",
    "apprenticeshipfun": "Collaboration & Teamwork & Social Culture",
    "apprenticeshipvariation": "Professional Development and Continuous Learning",
    "apprenticeshiprespect": "Collaboration & Teamwork & Social Culture"
}

id_counts_df['category'] = id_counts_df['id'].map(id_to_category)
df = id_counts_df
print(df)

# Saving new df file as excel file
df.to_excel('Kununu_rating_ids_mapped.xlsx')

### End of the notebook
